# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "XXX系统后端 HTTP REST API",
        "title": "XXX SERVER API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/xxx-server/v1",
    "paths": {
...
        "/task": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台任务"
                ],
                "summary": "查询后台任务列表",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "ID，支持多值",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务大类型：enum(自动任务-cron,下载任务-download)",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务子类型：enum(短临降雨-short,7天预报-mid,专题图-factor,反演报告-deduce)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "行政区划编码",
                        "name": "adcode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "要素类型",
                        "name": "factor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务状态：enum(NotStarted,InProc,Done,Failed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "创建时间（起止时间格式：2022-07-20 10:00:00,2022-07-21 10:00:00），只传一个表示不限制另一个",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面大小，默认10",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.CronTasksResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台任务"
                ],
                "summary": "批量删除后台任务",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "ids",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/task/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台任务"
                ],
                "summary": "获取单个后台任务详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.CronTaskResponse"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台任务"
                ],
                "summary": "重启后台任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/task/{id}/download": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台任务"
                ],
                "summary": "申请下载任务数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "下载任务参数",
                        "name": "factorThemeReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.FactorThemeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.IdResponse"
                        }
                    }
                }
            }
        },
        "/task/{id}/zip": {
            "get": {
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "后台任务"
                ],
                "summary": "下载任务数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/xxx": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "XXX数据"
                ],
                "summary": "查询XXX数据列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面大小，默认10",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "XXX数据"
                ],
                "summary": "新增XXX数据",
                "parameters": [
                    {
                        "description": "数据的值",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewXxxData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.IdResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "XXX数据"
                ],
                "summary": "批量删除XXX数据",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "ids",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/xxx/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "XXX数据"
                ],
                "summary": "获取单个XXX数据详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.NewXxxData": {
            "type": "object",
            "properties": {
                "district": {
                    "description": "地区",
                    "type": "string"
                },
                "path": {
                    "description": "路径",
                    "type": "string"
                }
            }
        },
        "entity.CronTask": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dir": {
                    "description": "任务输出目录",
                    "type": "string"
                },
                "end_at": {
                    "description": "结束时间",
                    "type": "string"
                },
                "err_log": {
                    "description": "任务错误日志",
                    "type": "string"
                },
                "ext": {
                    "description": "任务扩展信息",
                    "type": "object"
                },
                "ext_id": {
                    "description": "关联的外部任务/流程ID",
                    "type": "integer"
                },
                "genre": {
                    "description": "任务大类型 enum(cron,download)",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "任务名称",
                    "type": "string"
                },
                "params": {
                    "description": "任务参数",
                    "type": "object"
                },
                "progress": {
                    "description": "进度",
                    "type": "number"
                },
                "src_id": {
                    "description": "源任务ID",
                    "type": "integer"
                },
                "start_at": {
                    "description": "开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "任务状态 enum(NotStarted,InProc,Done,Failed)",
                    "type": "string"
                },
                "type": {
                    "description": "任务子类型 enum(short,mid,factor,deduce)",
                    "type": "string"
                }
            }
        },
        "entity.FactorThemeReq": {
            "type": "object",
            "properties": {
                "adcode": {
                    "description": "行政区划编码",
                    "type": "string"
                },
                "district": {
                    "description": "行政区划名称",
                    "type": "string"
                },
                "district_level": {
                    "description": "行政区划级别，省级（province）或市级（city）",
                    "type": "string"
                },
                "report_time": {
                    "description": "起报时刻（无需传入）",
                    "type": "string"
                },
                "themes": {
                    "description": "需要下载的要素列表",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "factor": {
                                "description": "XXX要素, enum(降雨,气压,温度,湿度,风场)",
                                "type": "string"
                            },
                            "title": {
                                "description": "专题图标题",
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "resp.CronTaskResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "$ref": "#/definitions/entity.CronTask"
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        },
        "resp.CronTasksResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "list": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CronTask"
                            }
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        },
        "resp.IdData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "resp.IdResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "$ref": "#/definitions/resp.IdData"
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        },
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code",
                    "type": "integer",
                    "enum": [
                        200,
                        400,
                        401,
                        500
                    ]
                },
                "data": {
                    "description": "data",
                    "type": "object"
                },
                "msg": {
                    "description": "message",
                    "type": "string"
                }
            }
        },
...
    }
}